Half precision.

+++ experiments/a_half.py
@@ -1,6 +1,6 @@
 from deeplabv3 import DeepWV3PlusTH
 import torch
-model = DeepWV3PlusTH(num_classes=19).cuda()
+model = DeepWV3PlusTH(num_classes=19).cuda().half()
 model.eval()
 from time import time_ns
 
@@ -10,7 +10,7 @@ def wrapper(image):
         with torch.no_grad():
             img = torch.from_numpy(image).float().unsqueeze(0).permute(0, 3, 1, 2) / 255.
             # img should be in [0, 1] and of shape 1x3xHxW
-            logit, logit_ood = model(img.cuda())
+            logit, logit_ood = model(img.cuda().half())
         out = torch.nn.functional.softmax(logit_ood, dim=1)
         p1 = torch.logsumexp(logit, dim=1)  # ln hat_p(x|din)
         p2 = out[:, 1]  # p(~din|x)
